add frame independentness restructure files a little bit
Former-commit-id: c7599698293fffc2048da3fbaddb5e887f195ff5

+++ transition/transition/AnimatorNode.h
@@ -19,6 +19,6 @@ public:
 		return{ this };
 	}
 
-	virtual void update(float delta) = 0;
+	virtual void update(double delta) = 0;
 };
